# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lyra personal productivity API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /habits:
    get:
      operationId: listHabits
      tags: [habits]
      summary: List habits
      responses:
        "200":
          description: List of habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Habit"
    post:
      operationId: createHabit
      tags: [habits]
      summary: Create a habit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHabitBody"
      responses:
        "201":
          description: Created habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
  /habits/{id}:
    patch:
      operationId: updateHabit
      tags: [habits]
      summary: Update a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHabitBody"
      responses:
        "200":
          description: Updated habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
    delete:
      operationId: deleteHabit
      tags: [habits]
      summary: Delete a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /habits/{id}/log:
    post:
      operationId: logHabit
      tags: [habits]
      summary: Log a habit completion for a date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogHabitBody"
      responses:
        "200":
          description: Habit log entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HabitLog"
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List calendar events
      parameters:
        - name: start
          in: query
          schema:
            type: ["string", "null"]
          description: Start date filter (YYYY-MM-DD)
        - name: end
          in: query
          schema:
            type: ["string", "null"]
          description: End date filter (YYYY-MM-DD)
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a calendar event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
  /events/{id}:
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventBody"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── PRAYERS ─────────────────────────────────────────────────────────────────
...
  /focus:
    get:
      operationId: listFocusSessions
      tags: [focus]
      summary: List focus sessions
      parameters:
        - name: date
          in: query
          schema:
            type: ["string", "null"]
          description: Filter by date (YYYY-MM-DD)
      responses:
        "200":
          description: List of focus sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FocusSession"
    post:
      operationId: createFocusSession
      tags: [focus]
      summary: Create/log a focus session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFocusSessionBody"
      responses:
        "201":
          description: Created focus session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FocusSession"
  /focus/{id}:
    patch:
      operationId: updateFocusSession
      tags: [focus]
      summary: Update a focus session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFocusSessionBody"
      responses:
        "200":
          description: Updated focus session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FocusSession"
    delete:
      operationId: deleteFocusSession
      tags: [focus]
      summary: Delete a focus session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Habit:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: ["string", "null"]
        timeOfDay:
          type: string
          enum: [morning, afternoon, evening, anytime]
        type:
          type: string
          enum: [positive, negative]
        streak:
          type: integer
        isArchived:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - timeOfDay
        - type
        - streak
        - isArchived
        - createdAt
    CreateHabitBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: ["string", "null"]
        timeOfDay:
          type: string
          enum: [morning, afternoon, evening, anytime]
        type:
          type: string
          enum: [positive, negative]
      required:
        - name
    UpdateHabitBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        timeOfDay:
          type: ["string", "null"]
        isArchived:
          type: ["boolean", "null"]
    HabitLog:
      type: object
      properties:
        id:
          type: string
        habitId:
          type: string
        date:
          type: string
        status:
          type: string
          enum: [completed, skipped, missed]
        loggedAt:
          type: string
      required:
        - id
        - habitId
        - date
        - status
        - loggedAt
    LogHabitBody:
      type: object
      properties:
        date:
          type: string
          description: Date in YYYY-MM-DD format
        status:
          type: string
          enum: [completed, skipped, missed]
      required:
        - date
        - status
...
    CalendarEvent:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
        allDay:
          type: boolean
        category:
          type: ["string", "null"]
        source:
          type: string
          enum: [native, task, prayer]
        createdAt:
          type: string
      required:
        - id
        - title
        - startTime
        - allDay
        - source
        - createdAt
    CreateEventBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
        allDay:
          type: boolean
        category:
          type: ["string", "null"]
      required:
        - title
        - startTime
    UpdateEventBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        allDay:
          type: ["boolean", "null"]
        category:
          type: ["string", "null"]

    # ─── PRAYER SCHEMAS ─────────────────────────────────────────────────────────
...
    FocusSession:
      type: object
      properties:
        id:
          type: string
        taskId:
          type: ["string", "null"]
        taskTitle:
          type: ["string", "null"]
        durationMinutes:
          type: integer
        status:
          type: string
          enum: [completed, interrupted]
        startedAt:
          type: string
        endedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - id
        - durationMinutes
        - status
        - startedAt
    CreateFocusSessionBody:
      type: object
      properties:
        taskId:
          type: ["string", "null"]
        taskTitle:
          type: ["string", "null"]
        durationMinutes:
          type: integer
        status:
          type: string
          enum: [completed, interrupted]
        startedAt:
          type: string
        endedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - durationMinutes
        - status
        - startedAt
    UpdateFocusSessionBody:
      type: object
      properties:
        durationMinutes:
          type: ["integer", "null"]
        status:
          type: ["string", "null"]
        endedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...